Replace pusher debug with dry-run option (#384)
* Replace pusher debug with dry-run option

This is a bit more general, the pusher debug option was always a bit
artificial. This is intended as a proper way to prevent the extractor
from persisting anything, while still letting users see what the
extractor would do.

This should mostly not access destinations at all. The exception is if
CDF is configured as source, which is intentional.

* Fix test

* Fix handling of skipped dps

* Move dry run termination into pushers

* Fix influx tests

* Move dry run check in PushEvents in InfluxPusher

+++ Extractor/Deletes.cs
@@ -84,14 +84,14 @@ namespace Cognite.OpcUa
             if (deletedStates.Any())
             {
                 logger.LogInformation("Found {Del} stored nodes in {Tab} that no longer exist and will be marked as deleted", deletedStates.Count, tableName);
-                await stateStore.DeleteExtractionState(deletedStates.Select(d => new NodeExistsState(d.Id, time)), tableName, token);
+                if (!config.DryRun) await stateStore.DeleteExtractionState(deletedStates.Select(d => new NodeExistsState(d.Id, time)), tableName, token);
             }
 
             var newStates = states.Values.Where(s => !oldStates.ContainsKey(s.Id)).ToList();
             if (newStates.Any())
             {
                 logger.LogInformation("Found {New} new nodes in {Tab}, adding to state store...", newStates.Count, tableName);
-                await stateStore.StoreExtractionState(newStates, tableName, s => new BaseStorableState { Id = s.Id }, token);
+                if (!config.DryRun) await stateStore.StoreExtractionState(newStates, tableName, s => new BaseStorableState { Id = s.Id }, token);
             }
 
             return deletedStates.Select(d => d.Id);
